113 Break models out into submodule
