+++ bartercheckout/apps.py
@@ -3,3 +3,4 @@ from django.apps import AppConfig
 
 class BartercheckoutConfig(AppConfig):
     name = 'bartercheckout'
+    verbose_name = 'Barter Checkout'
