+++ bartercheckout/custom_errors.py
@@ -1,0 +1,6 @@
+class BalanceLimitError(Exception):
+    """Raise when account balance exceeds limits"""
+
+
+class AmountInputError(Exception):
+    """Raise when amount entered is invalid"""
