+++ bartercheckout/models/__init__.py
@@ -1,0 +1,2 @@
+from .barter_account import BarterAccount
+from .barter_event import BarterEvent
